requirements.txt
flask
scikit-learn
joblib
pandas
numpy
pymysql
python-dotenv
